Add Update operation and edit code

+++ ToDoConsoleAppWithADO.NET/Program.cs
@@ -1,4 +1,5 @@
-﻿using ToDoConsoleAppWithADO.NET.Models;
+﻿using System.Linq;
+using ToDoConsoleAppWithADO.NET.Models;
 
 const string server = "localhost\\SQLEXPRESS";
 const string database = "ToDoListDB";
@@ -27,7 +28,6 @@ while (true)
             Console.Clear();
 
             var tasks = repository.ReadTasks();
-
             foreach (var item in tasks)
             {
                 Console.WriteLine(item);
@@ -44,14 +44,14 @@ while (true)
             Console.Write("Enter Task: ");
             string tempTaskName = Console.ReadLine()!;
 
-            Console.Write("Priority: ");
+            Console.Write("Priority (ex. Low, High, Critical): ");
             string tempTaskPriority = Console.ReadLine()!;
 
-            Console.Write("Status: ");
+            Console.Write("Status (ex. Not Started, In Progress, Complete): ");
             string tempTaskStatus = Console.ReadLine()!;
 
-            if (string.IsNullOrEmpty(tempTaskName) || string.IsNullOrEmpty(tempTaskPriority) || string.IsNullOrEmpty(tempTaskStatus))
-                continue;
+            //if (string.IsNullOrEmpty(tempTaskName) || string.IsNullOrEmpty(tempTaskPriority) || string.IsNullOrEmpty(tempTaskStatus))
+            //    continue;
 
             ToDoTask task = new ToDoTask()
             {
@@ -72,8 +72,32 @@ while (true)
         case ConsoleKey.D3:
             Console.Clear();
 
+            Console.Write("What you want update? (Task, Priority, Status): ");
+            string tempUpdate = Console.ReadLine()!;
+            Console.Write("which one? (Id): ");
+            int tempUpdateId = int.Parse(Console.ReadLine()!);
+            Console.Write("Enter text: ");
+            string tempUpdateText = Console.ReadLine()!;
 
+            bool updateSucces = false;
+            if (tempUpdate.Contains("Task") || tempUpdate.Contains("Priority") || tempUpdate.Contains("Status"))
+            {
+                if (tempUpdate.Contains("Status"))
+                {
+                    tempUpdate = "TaskStatuses";
+                }
+                if (tempUpdateId is not 0 && !string.IsNullOrEmpty(tempUpdateText))
+                {
+                    updateSucces = repository.UpdateTask(tempUpdate, tempUpdateId, tempUpdateText);
+                }
+            }
 
+            if (updateSucces)
+            {
+                Console.ForegroundColor = ConsoleColor.Yellow;
+                Console.WriteLine("Task updated");
+                Console.ResetColor();
+            }
 
             break;
 
@@ -83,11 +107,18 @@ while (true)
             Console.Write("Enter Task ID: ");
             int tempTaskId = int.Parse(Console.ReadLine()!);
 
-            bool deleteSucces = repository.DeleteTask(tempTaskId);
-            Console.ForegroundColor = ConsoleColor.Yellow;
-            Console.WriteLine("Task deleted");
-            Console.ResetColor();
+            bool deleteSucces = false;
+            if (tempTaskId is not 0)
+            {
+                deleteSucces = repository.DeleteTask(tempTaskId);
+            }
 
+            if (deleteSucces)
+            {
+                Console.ForegroundColor = ConsoleColor.Yellow;
+                Console.WriteLine("Task deleted");
+                Console.ResetColor();
+            }
             break;
 
         case ConsoleKey.Escape:
